Simplified version that doesn't break the deploy.
