+++ website/modules/@apostrophecms/seo/views/tagManagerBody.html
@@ -1,6 +1,3 @@
 {% if data.gtmId %}
-<!-- Google Tag Manager (noscript) -->
-<noscript><iframe src="https://www.googletagmanager.com/ns.html?id={{ data.gtmId }}"
-height="0" width="0" style="display:none;visibility:hidden" title="Google Tag Manager"></iframe></noscript>
-<!-- End Google Tag Manager (noscript) -->
+{% render '@apostrophecms/seo:gtmBody', data %}
 {% endif %}
