+++ website/modules/@apostrophecms/seo/views/tagManagerHead.html
@@ -1,9 +1,3 @@
 {% if data.gtmId %}
-<!-- Google Tag Manager -->
-<script>(function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
-new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
-j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
-'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
-})(window,document,'script','dataLayer','{{ data.gtmId }}');</script>
-<!-- End Google Tag Manager -->
+{% render '@apostrophecms/seo:gtmHead', data %}
 {% endif %}
