+++ website/package.json
@@ -40,7 +40,6 @@
   "dependencies": {
     "@apostrophecms/form": "^1.4.2",
     "@apostrophecms/import-export": "^3.2.0",
-    "@apostrophecms/seo": "^1.3.1",
     "@barba/core": "^2.10.3",
     "abort-controller": "^3.0.0",
     "apostrophe": "^4.17.0",
